NetApp cmode nfs: Fix QOS extra spec
This patch fixes the incorrect behavior where the NetApp cmode
nfs driver will choose to create a volume on a flexVol that has
a QOS policy group that matches the specified QOS policy group
intended for the cinder volume. The correct behavior is to ignore
the QOS policy group of the flexVol and instead assign the QOS policy
group to the newly created cinder volume.

Change-Id: I45d905da2a9a07b3ae8c00a225ab3b7f7ceb12d8
Closes-Bug: #1288283